requirements.txt
Django==2.1.5
django-ipware==2.1.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
requests==2.21.0